Reorganized utils
Split the big utils.py file into 3 files that are in the new utils folder.
